requirements.txt
beautifulsoup4
django-allauth
crispy_bootstrap5
django-crispy_forms
django-extensions
django-markdownx
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
